test: split test cases into respective files and proper request structure
